Fix editing databases for all types - some code duplication, but the code is readable
